requirements.txt
# Video/Audio Production System - Complete Dependencies
# ======================================================

# Core image processing
Pillow>=10.0.0

# Text-to-speech (neural voices)
edge-tts>=7.2.3

# Numerical operations (fast blending)
numpy>=1.24.0

# FFmpeg binaries (video encoding)
imageio-ffmpeg>=0.4.9

# Video editing library
moviepy>=2.1.1

# YAML parsing (input files)
PyYAML>=6.0

# HTTP requests (document fetching)
requests>=2.31.0

# YouTube transcript fetching
youtube-transcript-api>=0.6.0

# YouTube API (for search - optional, requires API key)
google-api-python-client>=2.100.0

# Installation:
# pip install -r requirements.txt
#
# Minimal install (without YouTube features):
# pip install Pillow edge-tts numpy imageio-ffmpeg PyYAML requests

# AI-powered features (optional but recommended)
# Claude API for:
# - Enhanced narration generation
# - High-quality translation (28+ languages)
anthropic>=0.34.0

# Environment variables (CRITICAL - missing dependency)
python-dotenv>=1.0.0

# Alternative translation (free fallback) - REMOVED due to dependency conflicts
# googletrans==4.0.0-rc1  # Conflicts with httpx, use Claude API instead

# Web UI dependencies (FastAPI + HTMX + Alpine.js)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
